experimental/sk_agent/pyproject.toml
[project]
name = "sk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-identity>=1.21.0",
    "azure-search-documents>=11.5.2",
    "azure-storage-blob>=12.25.1",
    "jupyter>=1.1.1",
    "numpy>=2.2.4",
    "openai>=1.68.2",
    "semantic-kernel>=1.25.0",
]
